+++ test/libsolidity/syntaxTests/types/mapping/mapping_data_location_function_param_public.sol
@@ -2,4 +2,4 @@ contract c {
     function f3(mapping(uint => uint) memory) view public {}
 }
 // ----
-// TypeError: (29-57): Mapping types as arguments or return values can only be used in internal or library functions.
+// TypeError: (29-57): Mapping types for parameters or return variables can only be used in internal or library functions.